requirements.txt
typer
python-dotenv
youtube-transcript-api
google-generativeai